Move AudioBus to Conductor
Move AudioBus and all its dependencies to the Conductor.
Register Manager as an optional delegate to handle
incoming MIDI through AudioBus (AKMIDIListener Protocol).

Move MIDIChannel and Omni configuration to Conductor.
